refactor(llama): extract ProcessLifecycle, ModelLoader, and Tokenization from LlamaBackend
Split three self-contained concerns out of the 835-line LlamaBackend.swift
into sibling files. Generation loop, stateLock, cancellation, unloadModel,
stopGeneration, and all cleanup remain in LlamaBackend.swift — those are
PR B.

- LlamaBackendProcessLifecycle: refcounts llama_backend_init / llama_backend_free
  at process scope (was static members + static lock on LlamaBackend).
- LlamaModelLoader: owns the C-level model load — parameter setup,
  progress-callback ABI bridging, RAII pointer wrappers, and the
  loadSerializationLock. LlamaBackend holds a single instance and delegates
  loadModel's detached task body through it.
- LlamaTokenization: static helpers for tokenize / tokenToString taking
  vocab as an explicit parameter, so callers remain responsible for
  guarding the pointer's lifetime.

Pure refactor — zero behavior change. Progresses #407 (PR A of two).
PR B (generation driver extraction) to follow.
